Find free position revision
Instead of a get_free_role function, I set up checking for duplicates in do_faceoff.

It functions mostly like HQMs, except it doesn't allow 2 goalies before filling all roles (this could be accounted for if we want to be accurate for this).

Once all positions are filled, it does indeed allow duplicates.